refactor: ElitePage 폴더 변경 및 하위 컴포넌트 이름 변경
