+++ src/components/routes/elite/EliteList/css/SettingBar.module.css
